return type change (division) and test adjustment

+++ src/main/java/calculator/Calculator.java
@@ -14,7 +14,7 @@ public class Calculator {
         return a*b;
     }
 
-    public double divTwoNumbers(int a, int b) {
+    public int divTwoNumbers(int a, int b) {
         return a/b;
     }
 }
